+++ zeus-modules/zeus-module-scenario/src/main/java/ch/wellernet/zeus/modules/scenario/scheduling/MidnightTrigger.java
@@ -7,7 +7,6 @@ import static java.util.Calendar.DATE;
 import java.util.Calendar;
 
 import com.luckycatlabs.sunrisesunset.Zenith;
-import com.luckycatlabs.sunrisesunset.dto.Location;
 
 import lombok.Builder;
 
@@ -23,12 +22,10 @@ public class MidnightTrigger extends DayTimeTrigger {
 		final Calendar yesterday = (Calendar) at.clone();
 		yesterday.add(DATE, -1);
 		final Calendar calendar = Calendar.getInstance();
-		final Calendar sunset = getSunset(getLocation().getLatitude().doubleValue(),
-				getLocation().getLongitude().doubleValue(), yesterday.getTimeZone(), yesterday,
-				90. - getZenith().degrees().doubleValue());
-		final Calendar sunrise = getSunrise(getLocation().getLatitude().doubleValue(),
-				getLocation().getLongitude().doubleValue(), at.getTimeZone(), at,
-				90. - getZenith().degrees().doubleValue());
+		final Calendar sunset = getSunset(getLocation().getLatitude(), getLocation().getLongitude(),
+				yesterday.getTimeZone(), yesterday, 90. - getZenith().degrees().doubleValue());
+		final Calendar sunrise = getSunrise(getLocation().getLatitude(), getLocation().getLongitude(), at.getTimeZone(),
+				at, 90. - getZenith().degrees().doubleValue());
 		if (sunrise == null || sunset == null) {
 			return null;
 		} else {
